Refactored SslOptions to single method.
